pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
uuid = "^1.30"
anthropic = "^0.40.0"
aiofiles = "^24.1.0"
python-magic = "^0.4.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
